# src/mail_client_service/openapi.json
{"openapi":"3.1.0","info":{"title":"Mail Client Service","description":"A REST API service that implements mail client operations","version":"1.0.0"},"paths":{"/messages":{"get":{"tags":["messages"],"summary":"List Messages","description":"List messages from the mail client.\n\nReturns a list of messages with their details.","operationId":"list_messages_messages_get","parameters":[{"name":"max_results","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Maximum number of messages to return","default":10,"title":"Max Results"},"description":"Maximum number of messages to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageResponse"},"title":"Response List Messages Messages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/messages/{message_id}":{"get":{"tags":["messages"],"summary":"Get Message","description":"Get a specific message by its ID.\n\nReturns the message details if found, otherwise raises a 404 error.","operationId":"get_message_messages__message_id__get","parameters":[{"name":"message_id","in":"path","required":true,"schema":{"type":"string","description":"ID of the message to retrieve","title":"Message Id"},"description":"ID of the message to retrieve"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["messages"],"summary":"Delete Message","description":"Delete a specific message.\n\nReturns success status. Raises 404 if message is not found.","operationId":"delete_message_messages__message_id__delete","parameters":[{"name":"message_id","in":"path","required":true,"schema":{"type":"string","description":"ID of the message to delete","title":"Message Id"},"description":"ID of the message to delete"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/messages/{message_id}/read":{"post":{"tags":["messages"],"summary":"Mark As Read","description":"Mark a message as read.\n\nReturns success status. Raises 404 if message is not found.","operationId":"mark_as_read_messages__message_id__read_post","parameters":[{"name":"message_id","in":"path","required":true,"schema":{"type":"string","description":"ID of the message to mark as read","title":"Message Id"},"description":"ID of the message to mark as read"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageResponse":{"properties":{"id":{"type":"string","title":"Id","description":"Unique identifier of the message"},"from":{"type":"string","title":"From","description":"Sender's email address"},"to":{"type":"string","title":"To","description":"Recipient's email address"},"date":{"type":"string","title":"Date","description":"Date the message was sent"},"subject":{"type":"string","title":"Subject","description":"Subject line of the message"},"body":{"type":"string","title":"Body","description":"Body content of the message","default":""}},"type":"object","required":["id","from","to","date","subject"],"title":"MessageResponse","description":"Represents an email message response.","example":{"body":"This is a test message body","date":"2025-10-03","from":"sender@example.com","id":"msg_123","subject":"Test Message","to":"recipient@example.com"}},"SuccessResponse":{"properties":{"success":{"type":"boolean","title":"Success","description":"Whether the operation was successful"}},"type":"object","required":["success"],"title":"SuccessResponse","description":"Standard success response."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}